Use a single jni.h and use more recent headers clearly labelled as GPLv2 w/CPE
